Create list wrapper class for "/options/" calls
Options responses are not structured in uniform with the rest of the API. All others return with count and total, along with data. Since responses for options calls do not distinguish these variables, I created a new class that does not extend StandardListWrapper.

Additionally, Option objects are returned as a data array, thus it is parsed as an "Option[]" array first, and then convert to a List.